+++ src/Plugin/Field/FieldFormatter/KeyFigureCondensed.php
@@ -3,7 +3,6 @@
 namespace Drupal\ocha_key_figures\Plugin\Field\FieldFormatter;
 
 use Drupal\Core\Field\FieldItemListInterface;
-use Drupal\ocha_key_figures\Helpers\NumberFormatter;
 
 /**
  * Plugin implementation of the 'key_figure' formatter.
@@ -22,11 +21,6 @@ class KeyFigureCondensed extends KeyFigureBase {
    * {@inheritdoc}
    */
   public function viewElements(FieldItemListInterface $items, $langcode) {
-    $format = $this->getSetting('format');
-    $precision = $this->getSetting('precision');
-    $percentage_formatted = $this->getSetting('percentage');
-    $currency_symbol = $this->getSetting('currency_symbol');
-
     $theme_suggestions = implode('__', [
       $this->viewMode,
       $items->getEntity()->getEntityTypeId(),
@@ -65,23 +59,9 @@ class KeyFigureCondensed extends KeyFigureBase {
       }
 
       foreach ($figures as $figure) {
-        // Set currency prefix if data is financial.
-        if (isset($figure['value_type']) && $figure['value_type'] == 'amount') {
-          $fig['prefix'] = $fig['unit'] ?? 'USD';
-          if ($currency_symbol == 'yes') {
-            $fig['prefix'] = NumberFormatter::getCurrencySymbol($langcode, $fig['prefix']);
-          }
-        }
+        $this->addPrefixSuffix($figure, $langcode);
 
-        // Set percentage suffix if needed.
-        if (isset($figure['value_type']) && $figure['value_type'] == 'percentage') {
-          $figure['unit'] = $figure['unit'] ?? '%';
-          if ($percentage_formatted != 'yes') {
-            $figure['value'] /= 100;
-          }
-        }
-
-        $value = NumberFormatter::format($figure['value'], $langcode, $format, $precision, FALSE);
+        $value = $this->formatNumber($figure['value'], $langcode);
         $elements['#figures'][] = [
           '#theme' => 'ocha_key_figures_figure__' . $theme_suggestions,
           '#label' => $figure['name'],
@@ -110,11 +90,7 @@ class KeyFigureCondensed extends KeyFigureBase {
             $value = $data['value'];
             $unit = $data['unit'] ?? '';
 
-            if ($data['value_type'] == 'percentage') {
-              if ($percentage_formatted == 'no') {
-                $value /= 100;
-              }
-            }
+            $this->addPrefixSuffix($data, $langcode);
           }
           else {
             $value = (string) $this->t('N/A');
@@ -122,7 +98,7 @@ class KeyFigureCondensed extends KeyFigureBase {
         }
 
         if (isset($label, $value)) {
-          $value = NumberFormatter::format($value, $langcode, $format, $precision, FALSE);
+          $value = $this->formatNumber($value, $langcode);
           $elements['#figures'][$delta] = [
             '#theme' => 'ocha_key_figures_figure__' . $this->viewMode,
             '#label' => $label,
