+++ src/Plugin/Field/FieldFormatter/KeyFigureExtended.php
@@ -6,7 +6,6 @@ use Drupal\Core\Field\FieldDefinitionInterface;
 use Drupal\Core\Field\FieldItemListInterface;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\ocha_key_figures\Controller\OchaKeyFiguresController;
-use Drupal\ocha_key_figures\Helpers\NumberFormatter;
 
 /**
  * Plugin implementation of the 'key_figure' formatter.
@@ -95,11 +94,6 @@ class KeyFigureExtended extends KeyFigureBase {
    * {@inheritdoc}
    */
   public function viewElements(FieldItemListInterface $items, $langcode) {
-    $format = $this->getSetting('format');
-    $precision = $this->getSetting('precision');
-    $percentage_formatted = $this->getSetting('percentage');
-    $currency_symbol = $this->getSetting('currency_symbol');
-
     $theme_suggestions = implode('__', [
       $this->viewMode,
       $items->getEntity()->getEntityTypeId(),
@@ -173,7 +167,7 @@ class KeyFigureExtended extends KeyFigureBase {
       $allowed_figure_ids = array_flip(preg_split('/,\s*/', trim(strtolower($allowed_figure_ids))));
       foreach ($filtered_results as $id => $figure) {
         if (!isset($allowed_figure_ids[$figure['figure_id']])) {
-          unset($figures[$id]);
+          unset($filtered_results[$id]);
         }
       }
     }
@@ -185,31 +179,15 @@ class KeyFigureExtended extends KeyFigureBase {
     }
 
     foreach ($data as &$fig) {
-      // Set currency prefix if data is financial.
-      if (isset($fig['value_type']) && $fig['value_type'] == 'amount') {
-        $fig['prefix'] = $fig['unit'] ?? 'USD';
-        if ($currency_symbol == 'yes') {
-          $fig['prefix'] = NumberFormatter::getCurrencySymbol($langcode, $fig['prefix']);
-        }
-      }
-
-      // Set percentage suffix if needed.
-      if (isset($fig['value_type']) && $fig['value_type'] == 'percentage') {
-        $fig['suffix'] = $fig['unit'] ?? '%';
-        if ($percentage_formatted != 'yes') {
-          $fig['value'] /= 100;
-        }
-      }
+      $this->addPrefixSuffix($fig, $langcode);
     }
 
-    $json_ld = NULL;
     if ($this->getSetting('output_json_ld') == 'yes') {
-      $json_ld = $this->buildJsonLd($data, $first);
-      $elements['#jsonld'] = $json_ld;
+      $elements['#jsonld'] = $this->buildJsonLd($data, $first);
     }
 
     foreach ($data as $figure) {
-      $figure['value'] = NumberFormatter::format($figure['value'], $langcode, $format, $precision, FALSE);
+      $figure['value'] = $this->formatNumber($figure['value'], $langcode);
       $elements['#figures'][] = [
         '#theme' => 'ocha_key_figures_extended_figure__' . $theme_suggestions,
         '#figure' => $figure,
@@ -221,213 +199,6 @@ class KeyFigureExtended extends KeyFigureBase {
     }
 
     return $elements;
-  }
-
-  /**
-   * Build json-ld output
-   */
-  protected function buildJsonLd($data, $item) {
-    // Initialize data for JSON-LD.
-    $json_data = [];
-    foreach ($data as $row) {
-      $json_data[$row['name']] = $row['value'];
-      $country = $row['country'];
-    }
-
-    $name = t('@title of @country', [
-      '@title' => 'Key figures',
-      '@country' => $country,
-    ]);
-    $description = t('Easily discoverable topline numbers for humanitarian crises in @country', [
-      '@country' => $country,
-    ]);
-
-    $metadata = $this->getMetaDataByProvider($item->getFigureProvider());
-    $metadata += [
-      'name' => $name,
-      'short_name' => 'Key figures',
-      'spatialCoverage' => $country,
-      'description' => $description,
-      'temporalCoverage' => $item->getFigureYear() === FALSE ? '2000-01-01/..' : $item->getFigureYear(),
-    ];
-
-    return $this->addJsonLdData($metadata, $json_data);
-  }
-
-  /**
-   * Get meta data for json ld.
-   */
-  protected function getMetaDataByProvider(string $provider) {
-    switch ($provider) {
-      case 'oct':
-        return [
-          'publisher' => [
-            '@type' => 'Organization',
-            'sameAs' => 'https://ror.org/00aahzn97',
-            'name' => 'OCHA Contribution Tracking (OCT)',
-          ],
-          'creator' => [
-            '@type' => 'Organization',
-            'sameAs' => 'https://ror.org/00aahzn97',
-            'name' => 'OCHA Contribution Tracking (OCT)',
-          ],
-          'license' => [
-            '@type' => 'CreativeWork',
-            'name' => 'Creative Commons Attribution for Intergovernmental Organisations',
-            'url' => 'https://data.humdata.org/faqs/licenses#auto-faq-_Data_Licenses_Content_-Creative_Commons_Attribution_for_Intergovernmental_Organisations__CC_BY_IGO_-a',
-          ],
-        ];
-
-      case 'fts':
-        return [
-          'publisher' => [
-            '@type' => 'Organization',
-            'sameAs' => 'https://ror.org/00aahzn97',
-            'name' => 'OCHA Financial Tracking System (FTS)',
-          ],
-          'creator' => [
-            '@type' => 'Organization',
-            'sameAs' => 'https://ror.org/00aahzn97',
-            'name' => 'OCHA Financial Tracking System (FTS)',
-          ],
-          'license' => [
-            '@type' => 'CreativeWork',
-            'name' => 'Creative Commons Attribution for Intergovernmental Organisations',
-            'url' => 'https://data.humdata.org/faqs/licenses#auto-faq-_Data_Licenses_Content_-Creative_Commons_Attribution_for_Intergovernmental_Organisations__CC_BY_IGO_-a',
-          ],
-        ];
-
-      case 'rw-crisis':
-        return [
-          'publisher' => [
-            '@type' => 'Organization',
-            'name' => 'ReliefWeb',
-          ],
-          'creator' => [
-            '@type' => 'Organization',
-            'name' => 'ReliefWeb',
-          ],
-          'license' => [
-            '@type' => 'CreativeWork',
-            'name' => 'Creative Commons Attribution for Intergovernmental Organisations',
-            'url' => 'https://data.humdata.org/faqs/licenses#auto-faq-_Data_Licenses_Content_-Creative_Commons_Attribution_for_Intergovernmental_Organisations__CC_BY_IGO_-a',
-          ],
-        ];
-
-      case 'inform-risk':
-        return [
-          'publisher' => [
-            '@type' => 'Organization',
-            'name' => 'ACAPS',
-          ],
-          'creator' => [
-            '@type' => 'Organization',
-            'name' => 'ACAPS',
-          ],
-          'license' => [
-            '@type' => 'CreativeWork',
-            'name' => 'Creative Commons Attribution International',
-            'url' => 'https://data.humdata.org/faqs/licenses#auto-faq-_Data_Licenses_Content_-Creative_Commons_Attribution_International_CC_BY_-a',
-          ],
-        ];
-
-      case 'cbpf':
-      case 'cerf':
-        return [
-          'publisher' => [
-            '@type' => 'Organization',
-            'sameAs' => 'https://ror.org/00aahzn97',
-            'name' => 'United Nations Office for the Coordination of Humanitarian Affairs',
-          ],
-          'creator' => [
-            '@type' => 'Organization',
-            'sameAs' => 'https://ror.org/00aahzn97',
-            'name' => 'United Nations Office for the Coordination of Humanitarian Affairs',
-          ],
-          'license' => [
-            '@type' => 'CreativeWork',
-            'name' => 'Creative Commons Attribution International',
-            'url' => 'https://data.humdata.org/faqs/licenses#auto-faq-_Data_Licenses_Content_-Creative_Commons_Attribution_International_CC_BY_-a',
-          ],
-        ];
-
-    }
-
-    // Default to OCHA.
-    return [
-      'publisher' => [
-        '@type' => 'Organization',
-        'sameAs' => 'https://ror.org/00aahzn97',
-        'name' => 'United Nations Office for the Coordination of Humanitarian Affairs',
-      ],
-      'creator' => [
-        '@type' => 'Organization',
-        'sameAs' => 'https://ror.org/00aahzn97',
-        'name' => 'United Nations Office for the Coordination of Humanitarian Affairs',
-      ],
-      'license' => [
-        '@type' => 'CreativeWork',
-        'name' => 'Creative Commons Attribution International',
-        'url' => 'https://data.humdata.org/faqs/licenses#auto-faq-_Data_Licenses_Content_-Creative_Commons_Attribution_International_CC_BY_-a',
-      ],
-    ];
-  }
-
-  /**
-   * Add json ld data.
-   */
-  protected function addJsonLdData($metadata, $data) {
-    $label_column = [];
-    $value_column = [];
-    $keywords = [];
-
-    // Default keyword.
-    $keywords[] = 'ReliefWeb > Numbers > ' . $metadata['spatialCoverage'] . ' > ' . $metadata['short_name'];
-
-    foreach ($data as $label => $value) {
-      $label_column[] = [
-        'csvw:value' => $label,
-        'csvw:primaryKey' => $label,
-      ];
-
-      $value_column[] = [
-        'csvw:value' => $value,
-        'csvw:primaryKey' => $value,
-      ];
-
-      $keywords[] = 'ReliefWeb > Numbers > ' . $metadata['spatialCoverage'] . ' > ' . $metadata['short_name'] . ' > ' . $label;
-    }
-
-    $json_ld = [
-      '@context' => [
-        'https://schema.org',
-        [
-          'csvw' => 'http://www.w3.org/ns/csvw#',
-        ],
-      ],
-      '@type' => 'Dataset',
-      'keywords' => $keywords,
-      'isAccessibleForFree' => TRUE,
-      'mainEntity' => [
-        '@type' => 'csvw:Table',
-        'csvw:tableSchema' => [
-          'csvw:columns' => [
-            [
-              'csvw:name' => 'Label',
-              'csvw:datatype' => 'string',
-              'csvw:cells' => $label_column,
-            ],
-            [
-              'csvw:name' => 'Value',
-              'csvw:datatype' => 'number',
-              'csvw:cells' => $value_column,
-            ],
-          ],
-        ],
-      ],
-    ] + $metadata;
-
-    return json_encode($json_ld);
   }
 
 }
